media-source/MediaSource-algorithm-DurationChange_SameDuration.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Check "duration change" algorithm of MediaSource Object in case new duration is equal to old duration</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm': 'video/webm; codecs="vp8"',
                '/media/white.mp4' : 'video/mp4'};

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
//check the browser supports the MIME used in this test
function isTypedSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function test(file, mime) {
    async_test(function(t) {
        if(!isTypedSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            video.controller = new MediaController();
            var durationchangeNum = 0;

            mediaSource.addEventListener('sourceopen', function(e) {
                mediaSource.duration = 5;//change duration

                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updateend', function(e) {
                    mediaSource.duration = 5;//set new duratoin whose is equal to old duration
                });
                sourceBuffer.appendBuffer(data);
            });
            video.addEventListener('durationchange', function(e) {
                durationchangeNum++;

                if(durationchangeNum > 1) {
                    t.step(function() {
                        assert_equals(durationchangeNum, 2, 'One of two is fired by video element, another one is fired by MediaSource.');
                        assert_equals(video.controller.duration, 5);
                        assert_equals(mediaSource.duration, 5);
                    });
                    t.done();
                }
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'Check "duration change" algorithm of MediaSource Object in case new duration is equal to old duration for (' + mime + ').');
}
for(var file in contents) {
    test(file, contents[file]);
}
</script>
</body>
</html>
